requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==4.2.19
django-filter==24.3
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-django==4.10.0
sqlparse==0.5.3
tzdata==2025.1
